Code cleanup p1
