+++ frontend/src-tauri/src/analytics/analytics.rs
