+++ frontend/src-tauri/src/analytics/mod.rs
@@ -1,0 +1,5 @@
+pub mod analytics;
+pub mod commands;
+
+pub use analytics::*;
+// Don't re-export commands to avoid conflicts - lib.rs will import directly
