+++ frontend/src-tauri/src/api/commands.rs
@@ -1,0 +1,2 @@
+// Commands are now directly in api.rs to avoid duplication
+// This file exists to maintain module structure compatibility
