+++ frontend/src-tauri/src/api/mod.rs
@@ -1,0 +1,5 @@
+pub mod api;
+pub mod commands;
+
+pub use api::*;
+// Don't re-export commands to avoid conflicts - lib.rs will import directly
