+++ frontend/src-tauri/src/audio/stream.rs
@@ -1,0 +1,261 @@
+use std::sync::Arc;
+use anyhow::Result;
+use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
+use cpal::{Device, Stream, SupportedStreamConfig};
+use log::{error, info, warn};
+
+use super::core::{AudioDevice, DeviceType, get_device_and_config};
+use super::pipeline::AudioCapture;
+use super::recording_state::RecordingState;
+
+/// Simplified audio stream wrapper
+pub struct AudioStream {
+    device: Arc<AudioDevice>,
+    stream: Stream,
+    capture: AudioCapture,
+}
+
+impl AudioStream {
+    /// Create a new audio stream for the given device
+    pub async fn create(
+        device: Arc<AudioDevice>,
+        state: Arc<RecordingState>,
+    ) -> Result<Self> {
+        info!("Creating audio stream for device: {}", device.name);
+
+        // Get the underlying cpal device and config
+        let (cpal_device, config) = get_device_and_config(&device).await?;
+
+        info!("Audio config - Sample rate: {}, Channels: {}, Format: {:?}",
+              config.sample_rate().0, config.channels(), config.sample_format());
+
+        // Create audio capture processor
+        let capture = AudioCapture::new(
+            device.clone(),
+            state.clone(),
+            config.sample_rate().0,
+            config.channels(),
+        );
+
+        // Build the appropriate stream based on sample format
+        let stream = Self::build_stream(&cpal_device, &config, capture.clone())?;
+
+        // Start the stream
+        stream.play()?;
+        info!("Audio stream started for device: {}", device.name);
+
+        Ok(Self {
+            device,
+            stream,
+            capture,
+        })
+    }
+
+    /// Build stream based on sample format
+    fn build_stream(
+        device: &Device,
+        config: &SupportedStreamConfig,
+        capture: AudioCapture,
+    ) -> Result<Stream> {
+        let config_copy = config.clone();
+
+        let stream = match config.sample_format() {
+            cpal::SampleFormat::F32 => {
+                let capture_clone = capture.clone();
+                device.build_input_stream(
+                    &config_copy.into(),
+                    move |data: &[f32], _: &cpal::InputCallbackInfo| {
+                        capture.process_audio_data(data);
+                    },
+                    move |err| {
+                        capture_clone.handle_stream_error(err);
+                    },
+                    None,
+                )?
+            }
+            cpal::SampleFormat::I16 => {
+                let capture_clone = capture.clone();
+                device.build_input_stream(
+                    &config_copy.into(),
+                    move |data: &[i16], _: &cpal::InputCallbackInfo| {
+                        let f32_data: Vec<f32> = data.iter()
+                            .map(|&sample| sample as f32 / i16::MAX as f32)
+                            .collect();
+                        capture.process_audio_data(&f32_data);
+                    },
+                    move |err| {
+                        capture_clone.handle_stream_error(err);
+                    },
+                    None,
+                )?
+            }
+            cpal::SampleFormat::I32 => {
+                let capture_clone = capture.clone();
+                device.build_input_stream(
+                    &config_copy.into(),
+                    move |data: &[i32], _: &cpal::InputCallbackInfo| {
+                        let f32_data: Vec<f32> = data.iter()
+                            .map(|&sample| sample as f32 / i32::MAX as f32)
+                            .collect();
+                        capture.process_audio_data(&f32_data);
+                    },
+                    move |err| {
+                        capture_clone.handle_stream_error(err);
+                    },
+                    None,
+                )?
+            }
+            cpal::SampleFormat::I8 => {
+                let capture_clone = capture.clone();
+                device.build_input_stream(
+                    &config_copy.into(),
+                    move |data: &[i8], _: &cpal::InputCallbackInfo| {
+                        let f32_data: Vec<f32> = data.iter()
+                            .map(|&sample| sample as f32 / i8::MAX as f32)
+                            .collect();
+                        capture.process_audio_data(&f32_data);
+                    },
+                    move |err| {
+                        capture_clone.handle_stream_error(err);
+                    },
+                    None,
+                )?
+            }
+            _ => {
+                return Err(anyhow::anyhow!("Unsupported sample format: {:?}", config.sample_format()));
+            }
+        };
+
+        Ok(stream)
+    }
+
+    /// Get device info
+    pub fn device(&self) -> &AudioDevice {
+        &self.device
+    }
+
+    /// Stop the stream
+    pub fn stop(self) -> Result<()> {
+        info!("Stopping audio stream for device: {}", self.device.name);
+        drop(self.stream);
+        Ok(())
+    }
+}
+
+/// Audio stream manager for handling multiple streams
+pub struct AudioStreamManager {
+    microphone_stream: Option<AudioStream>,
+    system_stream: Option<AudioStream>,
+    state: Arc<RecordingState>,
+}
+
+impl AudioStreamManager {
+    pub fn new(state: Arc<RecordingState>) -> Self {
+        Self {
+            microphone_stream: None,
+            system_stream: None,
+            state,
+        }
+    }
+
+    /// Start audio streams for the given devices
+    pub async fn start_streams(
+        &mut self,
+        microphone_device: Option<Arc<AudioDevice>>,
+        system_device: Option<Arc<AudioDevice>>,
+    ) -> Result<()> {
+        info!("Starting audio streams");
+
+        // Start microphone stream
+        if let Some(mic_device) = microphone_device {
+            match AudioStream::create(mic_device.clone(), self.state.clone()).await {
+                Ok(stream) => {
+                    self.state.set_microphone_device(mic_device);
+                    self.microphone_stream = Some(stream);
+                    info!("Microphone stream started successfully");
+                }
+                Err(e) => {
+                    error!("Failed to create microphone stream: {}", e);
+                    return Err(e);
+                }
+            }
+        }
+
+        // Start system audio stream
+        if let Some(sys_device) = system_device {
+            match AudioStream::create(sys_device.clone(), self.state.clone()).await {
+                Ok(stream) => {
+                    self.state.set_system_device(sys_device);
+                    self.system_stream = Some(stream);
+                    info!("System audio stream started successfully");
+                }
+                Err(e) => {
+                    warn!("Failed to create system audio stream: {}", e);
+                    // Don't fail if only system audio fails
+                }
+            }
+        }
+
+        // Ensure at least one stream was created
+        if self.microphone_stream.is_none() && self.system_stream.is_none() {
+            return Err(anyhow::anyhow!("No audio streams could be created"));
+        }
+
+        Ok(())
+    }
+
+    /// Stop all audio streams
+    pub fn stop_streams(&mut self) -> Result<()> {
+        info!("Stopping all audio streams");
+
+        let mut errors = Vec::new();
+
+        // Stop microphone stream
+        if let Some(mic_stream) = self.microphone_stream.take() {
+            if let Err(e) = mic_stream.stop() {
+                error!("Failed to stop microphone stream: {}", e);
+                errors.push(e);
+            }
+        }
+
+        // Stop system stream
+        if let Some(sys_stream) = self.system_stream.take() {
+            if let Err(e) = sys_stream.stop() {
+                error!("Failed to stop system stream: {}", e);
+                errors.push(e);
+            }
+        }
+
+        if !errors.is_empty() {
+            Err(anyhow::anyhow!("Failed to stop some streams: {:?}", errors))
+        } else {
+            info!("All audio streams stopped successfully");
+            Ok(())
+        }
+    }
+
+    /// Get stream count
+    pub fn active_stream_count(&self) -> usize {
+        let mut count = 0;
+        if self.microphone_stream.is_some() {
+            count += 1;
+        }
+        if self.system_stream.is_some() {
+            count += 1;
+        }
+        count
+    }
+
+    /// Check if any streams are active
+    pub fn has_active_streams(&self) -> bool {
+        self.microphone_stream.is_some() || self.system_stream.is_some()
+    }
+}
+
+impl Drop for AudioStreamManager {
+    fn drop(&mut self) {
+        if let Err(e) = self.stop_streams() {
+            error!("Error stopping streams during drop: {}", e);
+        }
+    }
+}
