+++ frontend/src-tauri/src/console_utils/commands.rs
@@ -1,0 +1,2 @@
+// Note: Tauri commands are defined in console_utils.rs to avoid duplicates
+// This file can be used for other command utilities if needed in the future
