+++ frontend/src-tauri/src/console_utils/console_utils.rs
