+++ frontend/src-tauri/src/ollama/commands.rs
@@ -1,0 +1,1 @@
+// Commands are now directly in ollama.rs to avoid duplication
