+++ frontend/src-tauri/src/openrouter/mod.rs
@@ -1,0 +1,5 @@
+pub mod openrouter;
+pub mod commands;
+
+pub use openrouter::*;
+// Don't re-export commands to avoid conflicts - lib.rs will import directly
