+++ frontend/src-tauri/src/whisper_engine/mod.rs
@@ -1,0 +1,5 @@
+pub mod whisper_engine;
+pub mod commands;
+
+pub use whisper_engine::*;
+pub use commands::*;
